# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Project Hub API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{id}/documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List documents for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: createDocument
      tags: [documents]
      summary: Upload/create a document for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDocumentBody"
      responses:
        "201":
          description: Created document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /documents/{id}:
    get:
      operationId: getDocument
      tags: [documents]
      summary: Get a document by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "404":
          description: Not found
    patch:
      operationId: updateDocument
      tags: [documents]
      summary: Update a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDocumentBody"
      responses:
        "200":
          description: Updated document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /documents/{id}/versions:
    get:
      operationId: listDocumentVersions
      tags: [documents]
      summary: List versions of a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of document versions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DocumentVersion"
    post:
      operationId: addDocumentVersion
      tags: [documents]
      summary: Add a new version to a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDocumentVersionBody"
      responses:
        "201":
          description: Created document version
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentVersion"
  /projects/{id}/issues:
    get:
      operationId: listIssues
      tags: [issues]
      summary: List issues for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of issues
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Issue"
    post:
      operationId: createIssue
      tags: [issues]
      summary: Create an issue for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIssueBody"
      responses:
        "201":
          description: Created issue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Issue"
  /issues/{id}:
    get:
      operationId: getIssue
      tags: [issues]
      summary: Get an issue by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Issue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Issue"
        "404":
          description: Not found
    patch:
      operationId: updateIssue
      tags: [issues]
      summary: Update an issue
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateIssueBody"
      responses:
        "200":
          description: Updated issue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Issue"
    delete:
      operationId: deleteIssue
      tags: [issues]
      summary: Delete an issue
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /projects/{id}/messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: List messages for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: createMessage
      tags: [messages]
      summary: Send a message in a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMessageBody"
      responses:
        "201":
          description: Created message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /messages/{id}:
    patch:
      operationId: updateMessage
      tags: [messages]
      summary: Update a message
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMessageBody"
      responses:
        "200":
          description: Updated message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
    delete:
      operationId: deleteMessage
      tags: [messages]
      summary: Delete a message
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /escalation-rules:
    get:
      operationId: listEscalationRules
      tags: [escalation]
      summary: List escalation rules
      parameters:
        - name: projectId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of escalation rules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EscalationRule"
    post:
      operationId: createEscalationRule
      tags: [escalation]
      summary: Create an escalation rule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEscalationRuleBody"
      responses:
        "201":
          description: Created escalation rule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EscalationRule"
  /escalation-rules/{id}:
    patch:
      operationId: updateEscalationRule
      tags: [escalation]
      summary: Update an escalation rule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEscalationRuleBody"
      responses:
        "200":
          description: Updated escalation rule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EscalationRule"
    delete:
      operationId: deleteEscalationRule
      tags: [escalation]
      summary: Delete an escalation rule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Document:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        stage:
          type: ["string", "null"]
        name:
          type: string
        fileUrl:
          type: ["string", "null"]
        fileType:
          type: ["string", "null"]
        fileSize:
          type: ["integer", "null"]
        version:
          type: integer
        uploadedBy:
          type: ["integer", "null"]
        uploadedAt:
          type: string
        approvalStatus:
          type: string
        approvedBy:
          type: ["integer", "null"]
        approvedAt:
          type: ["string", "null"]
        accessLevel:
          type: string
        tags:
          type: array
          items:
            type: string
        description:
          type: string
        createdAt:
          type: string
      required:
        - id
        - projectId
        - name
        - version
        - uploadedAt
        - approvalStatus
        - accessLevel
        - tags
        - createdAt
    CreateDocumentBody:
      type: object
      properties:
        name:
          type: string
        stage:
          type: string
        fileUrl:
          type: string
        fileType:
          type: string
        fileSize:
          type: integer
        uploadedBy:
          type: integer
        accessLevel:
          type: string
        tags:
          type: array
          items:
            type: string
        description:
          type: string
      required:
        - name
    UpdateDocumentBody:
      type: object
      properties:
        name:
          type: string
        stage:
          type: string
        fileUrl:
          type: string
        fileType:
          type: string
        fileSize:
          type: integer
        approvalStatus:
          type: string
        approvedBy:
          type: integer
        accessLevel:
          type: string
        tags:
          type: array
          items:
            type: string
        description:
          type: string
    DocumentVersion:
      type: object
      properties:
        id:
          type: integer
        documentId:
          type: integer
        version:
          type: integer
        fileUrl:
          type: ["string", "null"]
        uploadedBy:
          type: ["integer", "null"]
        uploadedAt:
          type: string
        notes:
          type: string
        createdAt:
          type: string
      required:
        - id
        - documentId
        - version
        - uploadedAt
        - createdAt
    CreateDocumentVersionBody:
      type: object
      properties:
        fileUrl:
          type: string
        uploadedBy:
          type: integer
        notes:
          type: string
    Issue:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        taskId:
          type: ["integer", "null"]
        milestoneId:
          type: ["integer", "null"]
        title:
          type: string
        description:
          type: string
        dependencyType:
          type: ["string", "null"]
        blockingOwnerId:
          type: ["integer", "null"]
        blockingDept:
          type: ["string", "null"]
        originalDeadline:
          type: ["string", "null"]
        proposedRevisedDeadline:
          type: ["string", "null"]
        status:
          type: string
          description: "Status: open, pending, resolved"
        raisedBy:
          type: ["integer", "null"]
        resolvedAt:
          type: ["string", "null"]
        resolutionNotes:
          type: string
        createdAt:
          type: string
      required:
        - id
        - projectId
        - title
        - status
        - createdAt
    CreateIssueBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        taskId:
          type: integer
        milestoneId:
          type: integer
        dependencyType:
          type: string
        blockingOwnerId:
          type: integer
        blockingDept:
          type: string
        originalDeadline:
          type: string
        proposedRevisedDeadline:
          type: string
        raisedBy:
          type: integer
      required:
        - title
    UpdateIssueBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        dependencyType:
          type: string
        blockingOwnerId:
          type: integer
        blockingDept:
          type: string
        originalDeadline:
          type: string
        proposedRevisedDeadline:
          type: string
        status:
          type: string
        resolutionNotes:
          type: string
    Message:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        taskId:
          type: ["integer", "null"]
        milestoneId:
          type: ["integer", "null"]
        senderId:
          type: integer
        body:
          type: string
        attachments:
          type: array
          items:
            type: object
        taggedUserIds:
          type: array
          items:
            type: integer
        threadParentId:
          type: ["integer", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - projectId
        - senderId
        - body
        - attachments
        - taggedUserIds
        - createdAt
        - updatedAt
    CreateMessageBody:
      type: object
      properties:
        senderId:
          type: integer
        body:
          type: string
        taskId:
          type: integer
        milestoneId:
          type: integer
        attachments:
          type: array
          items:
            type: object
        taggedUserIds:
          type: array
          items:
            type: integer
        threadParentId:
          type: integer
      required:
        - senderId
        - body
    UpdateMessageBody:
      type: object
      properties:
        body:
          type: string
        attachments:
          type: array
          items:
            type: object
        taggedUserIds:
          type: array
          items:
            type: integer
...
    EscalationRule:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: ["integer", "null"]
        triggerType:
          type: string
          description: "Trigger: rag_red, delay_days, budget_variance_pct"
        thresholdValue:
          type: number
        notifyUserIds:
          type: array
          items:
            type: integer
        isActive:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - triggerType
        - thresholdValue
        - notifyUserIds
        - isActive
        - createdAt
    CreateEscalationRuleBody:
      type: object
      properties:
        projectId:
          type: integer
        triggerType:
          type: string
        thresholdValue:
          type: number
        notifyUserIds:
          type: array
          items:
            type: integer
        isActive:
          type: boolean
      required:
        - triggerType
    UpdateEscalationRuleBody:
      type: object
      properties:
        triggerType:
          type: string
        thresholdValue:
          type: number
        notifyUserIds:
          type: array
          items:
            type: integer
        isActive:
          type: boolean
...